src/main/java/algorithm/sorting/InsetionSortRunningTime.java
package algorithm.sorting; 
import java.io.*;
import java.util.*;
import java.text.*;
import java.math.*;
import java.util.regex.*;

public class InsetionSortRunningTime {

    static int runningTime(int[] arr) {
        // Complete this function
        int count = 0 ; 
        for(int i = 1; i < arr.length; i++){
            int value = arr[i];
            int j = i - 1;
            while(j >= 0 && arr[j] > value){
                arr[j + 1] = arr[j];
                count ++ ; 
                j = j - 1;
            }
            arr[j + 1] = value;
        }
        return count ; 

        // printArray(A);
        // return A ; 

    }

    public static void main(String[] args) {
        Scanner in = new Scanner(System.in);
        int n = in.nextInt();
        int[] arr = new int[n];
        for(int arr_i = 0; arr_i < n; arr_i++){
            arr[arr_i] = in.nextInt();
        }
        int result = runningTime(arr);
        System.out.println(result);
        in.close();
    }
}
